test(e2e): expand harness-agent — absorb 8 standalone probes
Absorbs 8 pure-store probes into harness-agent.mjs as additional cases,
reducing electron launch count by 8.

Probes absorbed (file -> case name):
- probe-e2e-streaming-journey-switch.mjs -> streaming-journey-switch
- probe-e2e-streaming-journey-parallel.mjs -> streaming-journey-parallel
- probe-e2e-streaming-journey-queue-clear.mjs -> streaming-journey-queue-clear
- probe-e2e-streaming-journey-esc-interrupt.mjs -> streaming-journey-esc-interrupt
- probe-e2e-msg-queue.mjs -> msg-queue
- probe-e2e-esc-interrupt.mjs -> esc-interrupt
- probe-e2e-composer-morph-mention.mjs -> composer-morph-mention
- probe-e2e-user-block-hover-menu.mjs -> user-block-hover-menu

composer-morph-mention monkey-patches the `files:list` IPC handler on
main and uses registerDispose() to remove it before the next case runs.
Cases that assert on English UI strings (Stop button name, queued chip,
aria-labels) defensively force i18n=en at start since prior runs may
leave zh persisted in app_state.

Probe count: 64 -> 56.
